# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ravyu Business Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/onboarding:
    put:
      operationId: completeOnboarding
      tags: [user]
      summary: Save onboarding wizard data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/change-password:
    put:
      operationId: changePassword
      tags: [user]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
        "400":
          description: Wrong current password
...
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: List stored reviews
      parameters:
        - in: query
          name: rating
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResult"
  /reviews/scrape:
    post:
      operationId: scrapeReviews
      tags: [reviews]
      summary: Scrape reviews from Google Maps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeInput"
      responses:
        "200":
          description: Scrape complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResult"
        "400":
          description: Invalid URL
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a stored review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── AI ────────────────────────────────────────────────────────────────────
...
  /public/business/{slug}:
    get:
      operationId: getPublicBusiness
      tags: [public]
      summary: Public business landing page data
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicBusinessData"
        "404":
          description: Not found
  /public/business/{slug}/track-view:
    post:
      operationId: trackPageView
      tags: [public]
      summary: Track a page view on the public landing
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tracked
  /public/business/{slug}/track-review-click:
    post:
      operationId: trackReviewClick
      tags: [public]
      summary: Track a review CTA click on the public landing
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tracked
  /public/review/{slug}:
    get:
      operationId: getReviewCollectionPage
      tags: [public]
      summary: Review collection page data (Google review URL)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCollectionData"
        "404":
          description: Not found
  /public/review/{slug}/{token}:
    get:
      operationId: getTrackedReviewLink
      tags: [public]
      summary: Open tracked review link; mark request opened
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCollectionData"
        "404":
          description: Not found
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, username, profileComplete]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        challenges:
          type: ["array", "null"]
          items:
            type: string
        profileComplete:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
        phone:
          type: string
        company:
          type: string
    OnboardingInput:
      type: object
      required: [phone, company, businessType, industry, challenges]
      properties:
        phone:
          type: string
        company:
          type: string
        businessType:
          type: string
          enum: [single, multiple, franchise, agency]
        industry:
          type: string
          enum: [restaurant, healthcare, hospitality, retail, real_estate, professional, education, other]
        challenges:
          type: array
          items:
            type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # ── Reviews ───────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, author, rating, text, date, createdAt]
      properties:
        id:
          type: integer
        author:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
        date:
          type: string
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewListResult:
      type: object
      required: [reviews, total]
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
    ScrapeInput:
      type: object
      required: [mapsUrl]
      properties:
        mapsUrl:
          type: string
        maxReviews:
          type: integer
          default: 100
          maximum: 500
    ScrapeResult:
      type: object
      required: [imported, total]
      properties:
        imported:
          type: integer
        total:
          type: integer
        message:
          type: string

    # ── AI ────────────────────────────────────────────────────────────────────
...
    BusinessService:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
...
    PublicBusinessData:
      type: object
      required: [businessName, slug]
      properties:
        businessName:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        services:
          type: array
          items:
            $ref: "#/components/schemas/BusinessService"
        recentReviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        totalReviews:
          type: integer
        averageRating:
          type: number
        googleMapsUrl:
          type: ["string", "null"]
    ReviewCollectionData:
      type: object
      required: [businessName, googleReviewUrl]
      properties:
        businessName:
          type: string
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        googleReviewUrl:
          type: string
        customerName:
          type: ["string", "null"]
        personalized:
          type: boolean