# lib/api-zod/src/generated/types/scrapeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ravyu Business Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeInput {
  mapsUrl: string;
  /** @maximum 500 */
  maxReviews?: number;
}